# apps/web-v2/src/lib/generated/collections.openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "CurateDS.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
    "/collections": {
      "get": {
        "tags": [
          "CurateDS.Api"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CollectionResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "CurateDS.Api"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCollectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/collections/{collectionId}/items": {
      "get": {
        "tags": [
          "CurateDS.Api"
        ],
        "parameters": [
          {
            "name": "collectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "HasNoItemType",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "CreatedBeforeExclusive",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "ExactAttributeKey",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ExactAttributeValue",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "SearchText",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "LocationId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "TagIds",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "format": "uuid"
              }
            }
          },
          {
            "name": "AttributeFilters",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "SortBy",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "SortDirection",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "MinQuantity",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "MaxQuantity",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "CreatedAfter",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "CreatedBefore",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "HasNoLocation",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "HasNoTags",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "ItemTypeId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "TagMatchMode",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedItemsResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "CurateDS.Api"
        ],
        "parameters": [
          {
            "name": "collectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateItemRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemDetailResponse"
                }
              }
            }
          }
        }
      }
    },
    "/collections/{collectionId}/items/{itemId}": {
      "get": {
        "tags": [
          "CurateDS.Api"
        ],
        "parameters": [
          {
            "name": "collectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemDetailResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "CurateDS.Api"
        ],
        "parameters": [
          {
            "name": "collectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateItemRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemDetailResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "CurateDS.Api"
        ],
        "parameters": [
          {
            "name": "collectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CollectionResponse": {
        "required": [
          "id",
          "name",
          "createdUtc"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "createdUtc": {
            "type": "string",
            "format": "date-time"
          },
          "category": {
            "type": [
              "null",
              "string"
            ]
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "coverImageUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "color": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "CreateCollectionRequest": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "category": {
            "type": [
              "null",
              "string"
            ]
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "coverImageUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "color": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "PagedItemsResponse": {
        "required": [
          "items",
          "totalCount",
          "page",
          "pageSize",
          "totalPages"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItemSummaryResponse"
            }
          },
          "totalCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "page": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "pageSize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "totalPages": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "ItemSummaryResponse": {
        "required": [
          "id",
          "collectionId",
          "name",
          "description",
          "quantity",
          "locationId",
          "locationName",
          "tags",
          "attributeValueCount",
          "createdUtc",
          "updatedUtc",
          "primaryImageUrl"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "collectionId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "quantity": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "locationId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "locationName": {
            "type": [
              "null",
              "string"
            ]
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "attributeValueCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "createdUtc": {
            "type": "string",
            "format": "date-time"
          },
          "updatedUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "primaryImageUrl": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "CreateItemRequest": {
        "required": [
          "name",
          "description",
          "quantity",
          "locationId",
          "itemTypeId",
          "tagIds",
          "attributeValues"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "quantity": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "locationId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "itemTypeId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "tagIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "attributeValues": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateItemAttributeValueRequest"
            }
          }
        }
      },
      "CreateItemAttributeValueRequest": {
        "required": [
          "attributeDefinitionId",
          "value"
        ],
        "type": "object",
        "properties": {
          "attributeDefinitionId": {
            "type": "string",
            "format": "uuid"
          },
          "value": {
            "type": "string"
          }
        }
      },
      "ItemDetailResponse": {
        "required": [
          "id",
          "collectionId",
          "name",
          "description",
          "quantity",
          "locationId",
          "locationName",
          "itemTypeId",
          "tags",
          "createdUtc",
          "updatedUtc",
          "attributeValues",
          "mediaAssets"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "collectionId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "quantity": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "locationId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "locationName": {
            "type": [
              "null",
              "string"
            ]
          },
          "itemTypeId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagResponse"
            }
          },
          "createdUtc": {
            "type": "string",
            "format": "date-time"
          },
          "updatedUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "attributeValues": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItemAttributeValueResponse"
            }
          },
          "mediaAssets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MediaAssetResponse"
            }
          }
        }
      },
      "TagResponse": {
        "required": [
          "id",
          "name",
          "key",
          "createdUtc"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "key": {
            "type": "string"
          },
          "createdUtc": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ItemAttributeValueResponse": {
        "required": [
          "attributeDefinitionId",
          "attributeName",
          "attributeKey",
          "dataType",
          "value"
        ],
        "type": "object",
        "properties": {
          "attributeDefinitionId": {
            "type": "string",
            "format": "uuid"
          },
          "attributeName": {
            "type": "string"
          },
          "attributeKey": {
            "type": "string"
          },
          "dataType": {
            "$ref": "#/components/schemas/AttributeDataType"
          },
          "value": {
            "type": "string"
          }
        }
      },
      "AttributeDataType": {
        "enum": [
          "Text",
          "Number",
          "Decimal",
          "Boolean",
          "Date",
          "SingleSelect"
        ]
      },
      "MediaAssetResponse": {
        "required": [
          "id",
          "url",
          "contentType",
          "fileName",
          "sizeBytes",
          "isPrimary",
          "uploadedUtc"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "url": {
            "type": "string"
          },
          "contentType": {
            "type": "string"
          },
          "fileName": {
            "type": "string"
          },
          "sizeBytes": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "isPrimary": {
            "type": "boolean"
          },
          "uploadedUtc": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "UpdateItemRequest": {
        "required": [
          "name",
          "description",
          "quantity",
          "locationId",
          "itemTypeId",
          "tagIds",
          "attributeValues"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "quantity": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "locationId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "itemTypeId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "tagIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "attributeValues": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateItemAttributeValueRequest"
            }
          }
        }
      },
...
    }
  }
}